test tgeometry python gpu
