+++ python/test/core/test_core.py
@@ -27,14 +27,18 @@
 import open3d as o3d
 import numpy as np
 import pytest
-import core_test_utils
 
-if core_test_utils.torch_available():
+import sys
+import os
+sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/..")
+from open3d_test import list_devices, torch_available
+
+if torch_available():
     import torch
     import torch.utils.dlpack
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_creation(device):
     # Shape takes tuple, list or o3d.core.SizeVector
     t = o3d.core.Tensor.empty((2, 3), o3d.core.Dtype.Float32, device=device)
@@ -62,7 +66,7 @@ def test_creation(device):
 
 @pytest.mark.parametrize("shape", [(), (0,), (1,), (0, 2), (0, 0, 2),
                                    (2, 0, 3)])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_creation_special_shapes(shape, device):
     o3_t = o3d.core.Tensor.full(shape,
                                 3.14,
@@ -140,7 +144,7 @@ def test_size_vector():
         sv = o3d.core.SizeVector(["foo", "bar"])
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_tensor_constructor(device):
     dtype = o3d.core.Dtype.Int32
 
@@ -229,9 +233,9 @@ def test_tensor_to_numpy_scope():
     np.testing.assert_equal(dst_t, src_t)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_tensor_to_pytorch_scope(device):
-    if not core_test_utils.torch_available():
+    if not torch_available():
         return
 
     src_t = np.array([[10, 11, 12.], [13., 14., 15.]])
@@ -245,9 +249,9 @@ def test_tensor_to_pytorch_scope(device):
     np.testing.assert_equal(dst_t, src_t)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_tensor_from_to_pytorch(device):
-    if not core_test_utils.torch_available():
+    if not torch_available():
         return
 
     device_id = device.get_id()
@@ -282,7 +286,7 @@ def test_tensor_from_to_pytorch(device):
 
 
 def test_tensor_numpy_to_open3d_to_pytorch():
-    if not core_test_utils.torch_available():
+    if not torch_available():
         return
 
     # Numpy -> Open3D -> PyTorch all share the same memory
@@ -298,7 +302,7 @@ def test_tensor_numpy_to_open3d_to_pytorch():
     np.testing.assert_equal(r, c.cpu().numpy())
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_binary_ew_ops(device):
     a = o3d.core.Tensor(np.array([4, 6, 8, 10, 12, 14]), device=device)
     b = o3d.core.Tensor(np.array([2, 3, 4, 5, 6, 7]), device=device)
@@ -326,7 +330,7 @@ def test_binary_ew_ops(device):
     np.testing.assert_equal(a.cpu().numpy(), np.array([2, 2, 2, 2, 2, 2]))
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_to(device):
     a = o3d.core.Tensor(np.array([0.1, 1.2, 2.3, 3.4, 4.5,
                                   5.6]).astype(np.float32),
@@ -339,7 +343,7 @@ def test_to(device):
     assert b.device == a.device
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_unary_ew_ops(device):
     src_vals = np.array([0, 1, 2, 3, 4, 5]).astype(np.float32)
     src = o3d.core.Tensor(src_vals, device=device)
@@ -368,7 +372,7 @@ def test_unary_ew_ops(device):
                                atol=atol)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_getitem(device):
     np_t = np.array(range(24)).reshape((2, 3, 4))
     o3_t = o3d.core.Tensor(np_t, device=device)
@@ -392,7 +396,7 @@ def test_getitem(device):
                             np_t[0:2, 1:3, 0:4][0:1, 0:2, 2:3])
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_setitem(device):
     np_ref = np.array(range(24)).reshape((2, 3, 4))
     o3_ref = o3d.core.Tensor(np_ref, device=device)
@@ -494,7 +498,7 @@ def test_setitem(device):
     np.testing.assert_equal(o3_t.cpu().numpy(), np_t)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_cast_to_py_tensor(device):
     a = o3d.core.Tensor([1], device=device)
     b = o3d.core.Tensor([2], device=device)
@@ -506,7 +510,7 @@ def test_cast_to_py_tensor(device):
     "dim",
     [0, 1, 2, (), (0,), (1,), (2,), (0, 1), (0, 2), (1, 2), (0, 1, 2), None])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_sum(dim, keepdim, device):
     np_src = np.array(range(24)).reshape((2, 3, 4))
     o3_src = o3d.core.Tensor(np_src, device=device)
@@ -525,7 +529,7 @@ def test_reduction_sum(dim, keepdim, device):
     ((0, 2), (1)),
 ])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_special_shapes(shape_and_axis, keepdim, device):
     shape, axis = shape_and_axis
     np_src = np.array(np.random.rand(*shape))
@@ -541,7 +545,7 @@ def test_reduction_special_shapes(shape_and_axis, keepdim, device):
     "dim",
     [0, 1, 2, (), (0,), (1,), (2,), (0, 1), (0, 2), (1, 2), (0, 1, 2), None])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_mean(dim, keepdim, device):
     np_src = np.array(range(24)).reshape((2, 3, 4)).astype(np.float32)
     o3_src = o3d.core.Tensor(np_src, device=device)
@@ -555,7 +559,7 @@ def test_reduction_mean(dim, keepdim, device):
     "dim",
     [0, 1, 2, (), (0,), (1,), (2,), (0, 1), (0, 2), (1, 2), (0, 1, 2), None])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_prod(dim, keepdim, device):
     np_src = np.array(range(24)).reshape((2, 3, 4))
     o3_src = o3d.core.Tensor(np_src, device=device)
@@ -569,7 +573,7 @@ def test_reduction_prod(dim, keepdim, device):
     "dim",
     [0, 1, 2, (), (0,), (1,), (2,), (0, 1), (0, 2), (1, 2), (0, 1, 2), None])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_min(dim, keepdim, device):
     np_src = np.array(range(24))
     np.random.shuffle(np_src)
@@ -585,7 +589,7 @@ def test_reduction_min(dim, keepdim, device):
     "dim",
     [0, 1, 2, (), (0,), (1,), (2,), (0, 1), (0, 2), (1, 2), (0, 1, 2), None])
 @pytest.mark.parametrize("keepdim", [True, False])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_max(dim, keepdim, device):
     np_src = np.array(range(24))
     np.random.shuffle(np_src)
@@ -598,7 +602,7 @@ def test_reduction_max(dim, keepdim, device):
 
 
 @pytest.mark.parametrize("dim", [0, 1, 2, None])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_reduction_argmin_argmax(dim, device):
     np_src = np.array(range(24))
     np.random.shuffle(np_src)
@@ -614,7 +618,7 @@ def test_reduction_argmin_argmax(dim, device):
     np.testing.assert_allclose(o3_dst.cpu().numpy(), np_dst)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_advanced_index_get_mixed(device):
     np_src = np.array(range(24)).reshape((2, 3, 4))
     o3_src = o3d.core.Tensor(np_src, device=device)
@@ -640,7 +644,7 @@ def test_advanced_index_get_mixed(device):
     np.testing.assert_equal(o3_dst.cpu().numpy(), np_dst)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_advanced_index_set_mixed(device):
     np_src = np.array(range(24)).reshape((2, 3, 4))
     o3_src = o3d.core.Tensor(np_src, device=device)
@@ -663,9 +667,9 @@ def test_advanced_index_set_mixed(device):
     np.testing.assert_equal(o3_src.cpu().numpy(), np_src)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_tensor_from_to_pytorch(device):
-    if not core_test_utils.torch_available():
+    if not torch_available():
         return
 
 
@@ -675,7 +679,7 @@ def test_tensor_from_to_pytorch(device):
                                                        ("negative", "neg"),
                                                        ("exp", "exp"),
                                                        ("abs", "abs")])
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_unary_elementwise(np_func_name, o3_func_name, device):
     np_t = np.array([-3, -2, -1, 9, 1, 2, 3]).astype(np.float32)
     o3_t = o3d.core.Tensor(np_t, device=device)
@@ -693,7 +697,7 @@ def test_unary_elementwise(np_func_name, o3_func_name, device):
                                getattr(np, np_func_name)(np_t))
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_logical_ops(device):
     np_a = np.array([True, False, True, False])
     np_b = np.array([True, True, False, False])
@@ -713,7 +717,7 @@ def test_logical_ops(device):
     np.testing.assert_equal(o3_r.cpu().numpy(), np_r)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_comparision_ops(device):
     np_a = np.array([0, 1, -1])
     np_b = np.array([0, 0, 0])
@@ -728,7 +732,7 @@ def test_comparision_ops(device):
     np.testing.assert_equal((o3_a != o3_b).cpu().numpy(), np_a != np_b)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_non_zero(device):
     np_x = np.array([[3, 0, 0], [0, 4, 0], [5, 6, 0]])
     np_nonzero_tuple = np.nonzero(np_x)
@@ -738,7 +742,7 @@ def test_non_zero(device):
         np.testing.assert_equal(np_t, o3_t.cpu().numpy())
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_boolean_advanced_indexing(device):
     np_a = np.array([1, -1, -2, 3])
     o3_a = o3d.core.Tensor(np_a, device=device)
@@ -755,7 +759,7 @@ def test_boolean_advanced_indexing(device):
     np.testing.assert_equal(np_y, o3_y.cpu().numpy())
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_scalar_op(device):
     # +
     a = o3d.core.Tensor.ones((2, 3), o3d.core.Dtype.Float32, device=device)
@@ -994,7 +998,7 @@ def test_scalar_op(device):
     np.testing.assert_equal(a.cpu().numpy(), np.array([True, False, True]))
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_all_any(device):
     a = o3d.core.Tensor([False, True, True, True],
                         dtype=o3d.core.Dtype.Bool,
@@ -1013,7 +1017,7 @@ def test_all_any(device):
     assert not a.any()
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_allclose_isclose(device):
     a = o3d.core.Tensor([1, 2], device=device)
     b = o3d.core.Tensor([1, 3], device=device)
@@ -1038,7 +1042,7 @@ def test_allclose_isclose(device):
     assert not a.allclose(b)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_issame(device):
     dtype = o3d.core.Dtype.Float32
     a = o3d.core.Tensor.ones((2, 3), dtype, device=device)
@@ -1056,7 +1060,7 @@ def test_issame(device):
     assert d.issame(e)
 
 
-@pytest.mark.parametrize("device", core_test_utils.list_devices())
+@pytest.mark.parametrize("device", list_devices())
 def test_item(device):
     o3_t = o3d.core.Tensor.ones(
         (2, 3), dtype=o3d.core.Dtype.Float32, device=device) * 1.5
