+++ python/test/open3d_test.py
@@ -2,11 +2,46 @@ import os
 import sys
 import urllib.request
 import zipfile
+import numpy as np
+import pytest
 
 # Avoid pathlib to be compatible with Python 3.5+.
 __pwd = os.path.dirname(os.path.realpath(__file__))
 test_data_dir = os.path.join(__pwd, os.pardir, os.pardir, "examples",
                              "test_data")
+
+
+def torch_available():
+    try:
+        import torch
+        import torch.utils.dlpack
+    except ImportError:
+        return False
+    return True
+
+
+def list_devices():
+    """
+    If Open3D is built with CUDA support:
+    - If cuda device is available, returns [Device("CPU:0"), Device("CUDA:0")].
+    - If cuda device is not available, returns [Device("CPU:0")].
+
+    If Open3D is built without CUDA support:
+    - returns [Device("CPU:0")].
+    """
+    import open3d as o3d
+    devices = [o3d.core.Device("CPU:" + str(0))]
+    if torch_available() and o3d._build_config['BUILD_CUDA_MODULE']:
+        import torch
+        import torch.utils.dlpack
+        if (o3d.core.cuda.device_count() != torch.cuda.device_count()):
+            raise RuntimeError(
+                "o3d.core.cuda.device_count() != torch.cuda.device_count(), "
+                "{} != {}".format(o3d.core.cuda.device_count(),
+                                  torch.cuda.device_count()))
+    if o3d.core.cuda.device_count() > 0:
+        devices.append(o3d.core.Device("CUDA:0"))
+    return devices
 
 
 def download_fountain_dataset():
